[MRESOLVER-465] Make room for alternate version scheme
Refactor bits to make introduction of new scheme simpler.

---

https://issues.apache.org/jira/browse/MRESOLVER-465
