+++ maven-resolver-util/src/main/java/org/eclipse/aether/util/version/GenericVersionScheme.java
@@ -18,13 +18,7 @@
  */
 package org.eclipse.aether.util.version;
 
-import java.util.ArrayList;
-import java.util.Collection;
-
 import org.eclipse.aether.version.InvalidVersionSpecificationException;
-import org.eclipse.aether.version.VersionScheme;
-
-import static java.util.Objects.requireNonNull;
 
 /**
  * A version scheme using a generic version syntax and common sense sorting.
@@ -51,80 +45,10 @@ import static java.util.Objects.requireNonNull;
  * respectively, until the kind mismatch is resolved, e.g. "1-alpha" = "1.0.0-alpha" &lt; "1.0.1-ga" = "1.0.1".
  * </p>
  */
-public final class GenericVersionScheme implements VersionScheme {
-
-    /**
-     * Creates a new instance of the version scheme for parsing versions.
-     */
-    public GenericVersionScheme() {}
-
+public final class GenericVersionScheme extends VersionSchemeSupport {
     @Override
     public GenericVersion parseVersion(final String version) throws InvalidVersionSpecificationException {
         return new GenericVersion(version);
-    }
-
-    @Override
-    public GenericVersionRange parseVersionRange(final String range) throws InvalidVersionSpecificationException {
-        return new GenericVersionRange(range);
-    }
-
-    @Override
-    public GenericVersionConstraint parseVersionConstraint(final String constraint)
-            throws InvalidVersionSpecificationException {
-        String process = requireNonNull(constraint, "constraint cannot be null");
-
-        Collection<GenericVersionRange> ranges = new ArrayList<>();
-
-        while (process.startsWith("[") || process.startsWith("(")) {
-            int index1 = process.indexOf(')');
-            int index2 = process.indexOf(']');
-
-            int index = index2;
-            if (index2 < 0 || (index1 >= 0 && index1 < index2)) {
-                index = index1;
-            }
-
-            if (index < 0) {
-                throw new InvalidVersionSpecificationException(constraint, "Unbounded version range " + constraint);
-            }
-
-            GenericVersionRange range = parseVersionRange(process.substring(0, index + 1));
-            ranges.add(range);
-
-            process = process.substring(index + 1).trim();
-
-            if (process.startsWith(",")) {
-                process = process.substring(1).trim();
-            }
-        }
-
-        if (!process.isEmpty() && !ranges.isEmpty()) {
-            throw new InvalidVersionSpecificationException(
-                    constraint, "Invalid version range " + constraint + ", expected [ or ( but got " + process);
-        }
-
-        GenericVersionConstraint result;
-        if (ranges.isEmpty()) {
-            result = new GenericVersionConstraint(parseVersion(constraint));
-        } else {
-            result = new GenericVersionConstraint(UnionVersionRange.from(ranges));
-        }
-
-        return result;
-    }
-
-    @Override
-    public boolean equals(final Object obj) {
-        if (this == obj) {
-            return true;
-        }
-
-        return obj != null && getClass().equals(obj.getClass());
-    }
-
-    @Override
-    public int hashCode() {
-        return getClass().hashCode();
     }
 
     // CHECKSTYLE_OFF: LineLength
@@ -138,8 +62,9 @@ public final class GenericVersionScheme implements VersionScheme {
      */
     // CHECKSTYLE_ON: LineLength
     public static void main(String... args) {
-        System.out.println("Display parameters as parsed by Maven Resolver (in canonical form and as a list of tokens)"
-                + " and comparison result:");
+        System.out.println(
+                "Display parameters as parsed by Maven Resolver 'generic' scheme (in canonical form and as a list of tokens)"
+                        + " and comparison result:");
         if (args.length == 0) {
             return;
         }
